Cleans up and documents declarations
